Create readme.md
Polycystic ovary syndrome (PCOS) is a common endocrine disorder affecting ;10% to 15% of reproductive-aged women worldwide. Diagnosis requires two of the following: hyperandrogenism, oligo-ovulation or anovulation, and polycystic ovaries. In addition to reproductive dysfunction, many women with PCOS display metabolic abnormalities associated with hyperandrogenism. Recent studies have reported that the gut microbiome is altered in women with PCOS and rodent models of the disorder. However, it is unknown whether the gut microbiome plays a causal role in the development and pathology of PCOS. Previous studies in humans and in a letrozole-induced PCOS mouse model demonstrated significant changes in the gut microbial community. Here we describe a comparative analysis of deep sequencing gut microbiome results from 3 longitudinal studies performed using a letrozole-induce PCOS mouse model. The microbiome studies included data from gut microbiomes of letrozole and placebo mice all the same age range observed over 5 weeks. Letrozole treatment correlated with significant changes in abundance of the gut microbiome over time as seen in alpha diversity and beta diversity plot along with changes in abundance of specific bacteria as determined by Log2fold and Random Forest analyses. The bacteria that diverged between treatments included Akkermansia, Lactobacillus, Lachnospiraceae, Coprococcus, Bifidobacterium, Oscillospira and Erysipelotrichaceae, all of whom are suggested by other studies to play an important role in the gut microbiome. Furthermore, we extended our analysis to look at the significant bacteria individually to see how the abundance is affected in a letrozole treatment over time using a linear mixed model. This showed a high significance with most of the bacteria between treatment groups and over time, showing a better fit with a quadratic term which indicated a non-linear pattern that either peaked or troughed after a few weeks of treatment. This change corresponded with the start of puberty indicated the potential role hormonal changes might have on shaping the gut microbiome. Our results suggest modulation of gut microbiome could be a potential treatment option for PCOS with further investigation into the specific bacterial for potential novel probiotics.